Create a bag_of_words mapper that wraps ngrams to create a BOW representation
tft.ngrams returns ordered and possibly duplicated ngrams: e.g. the row "cats and dogs and cats" with ngram size of 1 produces ['cats', 'and', 'dogs', 'and', 'cats]. For many applications, you instead want a bag of words representation per row: {'cats', 'and', 'dogs'}.

This CL adds a bag_of_words mapper/FC that is a light wrapper around the existing ngrams and applies this de-duplication.

PiperOrigin-RevId: 252927517