Infer a Feature for RaggedTensor but don't allow a feature spec generation.
This CL allows RaggedTensors in when generating a Schema proto, and in that case generates a a Feature with the right dtype and no shape set.  It also adds the "ragged_tensor" tag.  This tag is detected in schema_as_feature_spec and results in an error.

The end to end behavior that is enabled by this CL is:

1) The preprocessing_fn can emit a RaggedTensor.
2) In this case, a Schema proto will be generated, including information such as vocabulary sizes and other metadata.
3) However TFTransformOutput.transformed_feature_spec will fail.
4) Similarly materialization will fail in the Transform TFX component.

PiperOrigin-RevId: 260528422