Save ConcreteFunction instead of tf.function to make sure the lookup table workaround doesn't break
tf.function contains lookup table creates a new lookup table each time the function is traced. The current workaround uses a ResourceTracker to catch, track and save the created lookup table. This relies on the assumption that the tf.function won't be retraced further. A upcoming change in saved_model will invalidate this assumption. Saving a concrete function guarantees that it won't be retraced.

PiperOrigin-RevId: 336422726