Fix the decoding issue of fingerprint_sort_fn.
The previous implementation converts a bytes to str (requiring a successful decode) and then back to bytes (for taking hash) which leads to decoding failures when there is any invalid utf. The original intention is to be defensive on the input but it seems upstream code has made sure the input is always bytes and therefore no need to make such conversions.

PiperOrigin-RevId: 413435428