Fix to only transform raw data when requested.
When read_raw_data_for_training is set to False when invoking the main
function, common.transform_data was being called on raw train and test
data anyway.  This fix moves the transformation to the block where
read_raw_data_for_training is True. The scenario here is the data has
already been preprocessed, and the user wishes to re-use that
preprocessed data.