Shape related code is simplified by using (1,).
Since this is tabular data we're dealing with, the code has been
updated to treat it as such.  The net result is simpler shape related
code.  Education-num is treated as dense here instead of sparse as it
was before.  It might be missing values in the data, so it might call
for some sort of imputation to be done.