Switching instance dict transformed data path to convert to pa.RecordBatches first.
This allows us to fully outsource unbatching and conversion from batched tensors to tfx_bsl's tensor_to_arrow. This slows down the conversion to instance dicts by ~2x as shown by the benchmarks, but allows us to get rid of a good chunk of the tricky conversion logic and will make adding support for new outputs easier. Given that this only affects `output_record_batches=False` path that is demonstrative and not performance-focused already, I think it's reasonable.

PiperOrigin-RevId: 502669178