Fix uses of the np.stack() family of functions under NumPy 1.25.
NumPy 1.25 release notes: https://github.com/numpy/numpy/releases/tag/v1.25.0

Per the release notes, in NumPy 1.25:

A sequence must now be passed into the stacking family of functions (stack, vstack, hstack, dstack and column_stack). (gh-23019)

This change fixes test failures where a non-sequence was passed to a stack function. The most common fix is to convert the input to a list explicitly before passing it to any stacking functions.
END_PUBLIC

If additional changes are needed, such as exporting the change as an upstream PR or adding it as a copybara patch, I'd very much appreciate it if the OWNERS would do so.

Tested:
    TAP train for global presubmit queue
    http://test/OCL:542645113:BASE:542596798:1687481512144:cee8989d

BEGIN_PUBLIC
Internal Code Change

PiperOrigin-RevId: 542826613